+++ NEWS.md
@@ -1,5 +1,17 @@
 * `NEWS.md` file tracks changes made to the package.
 
+`GARCOM 1.2.0 Date Sept 10 2020`
+- Minor addition of features with VCF counts: filter genes
+
+`GARCOM 1.1.0 Date Sept 09 2020`
+- Minor addition of features with VCF counts: keep_indiv and extract SNPs
+
+`GARCOM 1.0.0 Date Sept 08 2020`
+- Prepare for major release one
+
+`GARCOM 0.0.0.9005 Date Sept 08 2020`
+- Added features for VCF: two new functions vcf_counts_SNP_genecoords 2) vcf_counts_annot
+
 `GARCOM 0.0.0.9003 Date Aug 18 2020`
 - Added impute feature to the functions for missingness. We allow mean and median